requirements.txt
afmparser==1.0.3
alembic==1.16.4
aniso8601==10.0.1
antlr4-python3-runtime==4.13.1
astutils==0.0.6
attrs==25.3.0
Authlib==1.6.1
autoflake==2.3.1
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.1.0
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachelib==0.13.0
captcha==0.7.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
coverage==7.10.1
cryptography==45.0.5
dd==0.5.7
dnspython==2.7.0
docker==7.1.0
elastic-transport==8.17.1
elasticsearch==8.18.1
email_validator==2.2.0
Faker==37.4.2
flake8==7.3.0
Flake8-pyproject==1.2.3
flamapy==2.0.1
flamapy-bdd==2.0.1
flamapy-fm==2.0.1
flamapy-fw==2.0.1
flamapy-sat==2.0.1
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gevent==25.5.1
geventhttpclient==2.3.4
graphviz==0.21
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaitaistruct==0.10
locust==2.37.14
locust-cloud==1.26.3
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mistune==3.1.3
msgpack==1.1.1
msgspec==0.19.0
mypy_extensions==1.1.0
networkx==3.5
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.10
psutil==7.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydot==4.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-sat==0.1.8.dev17
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rq==2.4.1
selenium==4.34.2
selenium-wire==5.1.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
SQLAlchemy-Utils==0.41.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvlparser==2.0.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.1.1
zope.interface==7.2
zstandard==0.23.0
pre-commit==4.3.0
gpxpy==1.6.2
